Remove unused atom masks and make atom masks 64 bits

+++ src/KOKKOS/atom_vec_angle_kokkos.cpp
@@ -662,7 +662,7 @@ int AtomVecAngleKokkos::unpack_exchange_kokkos(DAT::tdual_double_2d_lr &k_buf, i
 
 /* ---------------------------------------------------------------------- */
 
-void AtomVecAngleKokkos::sync(ExecutionSpace space, unsigned int mask)
+void AtomVecAngleKokkos::sync(ExecutionSpace space, uint64_t mask)
 {
   if (space == Device) {
     if (mask & X_MASK) atomKK->k_x.sync_device();
@@ -744,7 +744,7 @@ void AtomVecAngleKokkos::sync(ExecutionSpace space, unsigned int mask)
 
 /* ---------------------------------------------------------------------- */
 
-void AtomVecAngleKokkos::sync_pinned(ExecutionSpace space, unsigned int mask, int async_flag)
+void AtomVecAngleKokkos::sync_pinned(ExecutionSpace space, uint64_t mask, int async_flag)
 {
   if (space == Device) {
     if ((mask & X_MASK) && atomKK->k_x.need_sync_device())
@@ -837,7 +837,7 @@ void AtomVecAngleKokkos::sync_pinned(ExecutionSpace space, unsigned int mask, in
 
 /* ---------------------------------------------------------------------- */
 
-void AtomVecAngleKokkos::modified(ExecutionSpace space, unsigned int mask)
+void AtomVecAngleKokkos::modified(ExecutionSpace space, uint64_t mask)
 {
   if (space == Device) {
     if (mask & X_MASK) atomKK->k_x.modify_device();
