+++ src/KOKKOS/atom_vec_bond_kokkos.h
@@ -50,9 +50,9 @@ class AtomVecBondKokkos : public AtomVecKokkos, public AtomVecBond {
                              ExecutionSpace space,
                              DAT::tdual_int_1d &k_indices) override;
 
-  void sync(ExecutionSpace space, unsigned int mask) override;
-  void modified(ExecutionSpace space, unsigned int mask) override;
-  void sync_pinned(ExecutionSpace space, unsigned int mask, int async_flag = 0) override;
+  void sync(ExecutionSpace space, uint64_t mask) override;
+  void modified(ExecutionSpace space, uint64_t mask) override;
+  void sync_pinned(ExecutionSpace space, uint64_t mask, int async_flag = 0) override;
 
  private:
   tagint *molecule;
