+++ src/KOKKOS/atom_vec_dpd_kokkos.h
@@ -59,9 +59,9 @@ class AtomVecDPDKokkos : public AtomVecKokkos, public AtomVecDPD {
                              ExecutionSpace space,
                              DAT::tdual_int_1d &k_indices) override;
 
-  void sync(ExecutionSpace space, unsigned int mask) override;
-  void modified(ExecutionSpace space, unsigned int mask) override;
-  void sync_pinned(ExecutionSpace space, unsigned int mask, int async_flag = 0) override;
+  void sync(ExecutionSpace space, uint64_t mask) override;
+  void modified(ExecutionSpace space, uint64_t mask) override;
+  void sync_pinned(ExecutionSpace space, uint64_t mask, int async_flag = 0) override;
 
   double *duChem;
