+++ src/KOKKOS/atom_vec_full_kokkos.cpp
@@ -895,7 +895,7 @@ int AtomVecFullKokkos::unpack_exchange_kokkos(DAT::tdual_double_2d_lr &k_buf, in
 
 /* ---------------------------------------------------------------------- */
 
-void AtomVecFullKokkos::sync(ExecutionSpace space, unsigned int mask)
+void AtomVecFullKokkos::sync(ExecutionSpace space, uint64_t mask)
 {
   if (space == Device) {
     if (mask & X_MASK) atomKK->k_x.sync_device();
@@ -1028,7 +1028,7 @@ void AtomVecFullKokkos::sync(ExecutionSpace space, unsigned int mask)
 
 /* ---------------------------------------------------------------------- */
 
-void AtomVecFullKokkos::sync_pinned(ExecutionSpace space, unsigned int mask, int async_flag)
+void AtomVecFullKokkos::sync_pinned(ExecutionSpace space, uint64_t mask, int async_flag)
 {
   if (space == Device) {
     if ((mask & X_MASK) && atomKK->k_x.need_sync_device())
@@ -1179,7 +1179,7 @@ void AtomVecFullKokkos::sync_pinned(ExecutionSpace space, unsigned int mask, int
 
 /* ---------------------------------------------------------------------- */
 
-void AtomVecFullKokkos::modified(ExecutionSpace space, unsigned int mask)
+void AtomVecFullKokkos::modified(ExecutionSpace space, uint64_t mask)
 {
   if (space == Device) {
     if (mask & X_MASK) atomKK->k_x.modify_device();
