+++ src/KOKKOS/atom_vec_hybrid_kokkos.h
@@ -96,9 +96,9 @@ class AtomVecHybridKokkos : public AtomVecKokkos, public AtomVecHybrid {
                              ExecutionSpace space,
                              DAT::tdual_int_1d &k_indices) override;
 
-  void sync(ExecutionSpace space, unsigned int mask) override;
-  void modified(ExecutionSpace space, unsigned int mask) override;
-  void sync_pinned(ExecutionSpace space, unsigned int mask, int async_flag = 0) override;
+  void sync(ExecutionSpace space, uint64_t mask) override;
+  void modified(ExecutionSpace space, uint64_t mask) override;
+  void sync_pinned(ExecutionSpace space, uint64_t mask, int async_flag = 0) override;
 
  private:
   DAT::t_tagint_1d d_tag;
